ヘッダーのタイトル名追加・特定のURLのみ登録するためバリデーションの変更・csvファイルのみ登録できるよう実装

+++ app/controllers/keywords_controller.rb
@@ -10,36 +10,29 @@ class KeywordsController < ApplicationController
 
   def create
     @keyword = Keyword.new(keyword_params)
-    # filedir  = "#{Rails.root}/tmp"
-    # @keywordから利用ファイル名の情報を変数filepathに入れる。
-    # filepath = "#{filedir}/#{filename}"
-    # 変数filenameのファイルを開く処理を変数fpに入れている。
-    #fp = open(filename,'r')
-    # 変数fpのファイルを開き、ファイルの行数を数え、その情報を変数line_countに入れている
-    # line_count = fp.read.count("\n")
-    # 行数が70を超えていた場合
-    # if line_count >= 75
-      # 行数が超えていた場合のメッセージ表示のためflash利用
-      # @key = flash[:@key] = "⚠︎利用ファイルのキーワード数を70以下にしてください。"
-      # render :new, status: :unprocessable_entity
-    uploaded_file = params[:keyword][:open_filename]
-    filedir  = "#{Rails.root}/tmp"
-    csv_data = uploaded_file.read
-    file_name = uploaded_file.original_filename
-    filepath = "#{filedir}/#{file_name}"
-    File.open(filepath, 'wb') do |file|
-      file.write(csv_data)
-    end
-    fp = open(filepath,'r')
-    line_count = fp.read.count("\n")
-    if line_count >= 51
-      @key = flash[:@key] = "⚠︎利用ファイルのキーワード数を50以下にしてください。"
-      File.delete(filepath)
-      render :new, status: :unprocessable_entity
-    elsif @keyword.save
-      KeywordJob.perform_async(file_name, @keyword.id)
-      redirect_to action: :index
+    if @keyword.open_filename.attached? && @keyword.open_filename.content_type == 'text/csv'
+      uploaded_file = params[:keyword][:open_filename]
+      filedir  = "#{Rails.root}/tmp"
+      csv_data = uploaded_file.read
+      file_name = uploaded_file.original_filename
+      filepath = "#{filedir}/#{file_name}"
+      File.open(filepath, 'wb') do |file|
+        file.write(csv_data)
+      end
+      fp = open(filepath,'r')
+      line_count = fp.read.count("\n")
+      if line_count >= 51
+        @key = flash[:@key] = "⚠︎利用ファイルのキーワード数を50以下にしてください。"
+        File.delete(filepath)
+        render :new, status: :unprocessable_entity
+      elsif @keyword.save
+        KeywordJob.perform_async(file_name, @keyword.id)
+        redirect_to action: :index
+      else
+        render :new, status: :unprocessable_entity
+      end
     else
+      @key = flash[:@key] = "⚠︎CSVファイルのみアップロード可能です"
       render :new, status: :unprocessable_entity
     end
   end
@@ -60,6 +53,9 @@ class KeywordsController < ApplicationController
     redirect_to action: :index
   end
 
+  def show
+
+  end
   def search
     @keywords = Keyword.page(params[:page])
     @keyword = Keyword.search(params[:keyword])
